+++ Kernel/Modules/AdminSystemConfiguration.pm
@@ -136,7 +136,8 @@ sub Run {
         # Check if setting is fixed but not yet deployed.
         for my $SettingName (@SettingNames) {
             my %Setting = $SysConfigObject->SettingGet(
-                Name => $SettingName,
+                Name           => $SettingName,
+                OverriddenInXML => 1,
             );
 
             my %EffectiveValueCheck = $SysConfigObject->SettingEffectiveValueCheck(
@@ -271,7 +272,8 @@ sub Run {
                 for my $SettingName ( sort @Result ) {
 
                     my %Setting = $SysConfigObject->SettingGet(
-                        Name => $SettingName,
+                        Name           => $SettingName,
+                        OverriddenInXML => 1,
                     );
 
                     $Setting{HTMLStrg} = $SysConfigObject->SettingRender(
@@ -345,7 +347,8 @@ sub Run {
             for my $SettingName ( sort @{$Favourites} ) {
 
                 my %Setting = $SysConfigObject->SettingGet(
-                    Name => $SettingName,
+                    Name           => $SettingName,
+                    OverriddenInXML => 1,
                 );
 
                 $Setting{HTMLStrg} = $SysConfigObject->SettingRender(
@@ -390,7 +393,8 @@ sub Run {
             my $SysConfigObject = $Kernel::OM->Get('Kernel::System::SysConfig');
 
             my %Setting = $SysConfigObject->SettingGet(
-                Name => $View,
+                Name           => $View,
+                OverriddenInXML => 1,
             );
 
             if (%Setting) {
@@ -433,7 +437,8 @@ sub Run {
             for my $Name ( sort @Names ) {
 
                 my %Setting = $SysConfigObject->SettingGet(
-                    Name => $Name,
+                    Name           => $Name,
+                    OverriddenInXML => 1,
                 );
 
                 if (%Setting) {
@@ -491,7 +496,8 @@ sub Run {
             for my $SettingName ( sort @Result ) {
 
                 my %Setting = $SysConfigObject->SettingGet(
-                    Name => $SettingName,
+                    Name           => $SettingName,
+                    OverriddenInXML => 1,
                 );
 
                 $Setting{HTMLStrg} = $SysConfigObject->SettingRender(